Add jsp view

+++ SpringBootBlog/src/test/java/blog/repository/ArticleRepositoryTest.java
@@ -7,9 +7,7 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import blog.repository.ArticleRepository;
 import blog.entity.Article;
@@ -28,7 +26,7 @@ public class ArticleRepositoryTest {
 	public void testFindById() {
 		Article article = new Article("1st", "abcdefg");
 		entityManager.persist(article);
-		Article a = articles.findById(1);
+		Article a = articles.findById((long) 1);
 		assertNotNull(a);
 	}
 }
